@@ -4,26 +4,28 @@
   "scripts": {
     "dev": "vite dev",
     "build": "vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "format": "prettier . --write"
   },
   "devDependencies": {
     "@fontsource/fira-mono": "^4.5.10",
     "@neoconfetti/svelte": "^1.0.0",
     "@sveltejs/adapter-auto": "^3.0.0",
     "@sveltejs/adapter-static": "^3.0.1",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/kit": "^2.20.7",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "@tailwindcss/typography": "^0.5.10",
     "autoprefixer": "^10.4.16",
     "daisyui": "^4.4.20",
     "postcss": "^8.4.32",
     "postcss-load-config": "^5.0.2",
-    "svelte": "^4.2.7",
+    "prettier": "^3.5.3",
+    "prettier-plugin-svelte": "^3.3.3",
+    "prettier-plugin-tailwindcss": "^0.6.11",
+    "svelte": "^5.28.2",
+    "svelte-preprocess": "^6.0.3",
     "tailwindcss": "^3.3.6",
-    "vite": "^5.0.3"
+    "vite": "^6.3.3"
   },
-  "type": "module",
-  "dependencies": {
-    "svelte-preprocess": "^5.1.3"
-  }
+  "type": "module"
 }
