@@ -1,13 +1,12 @@
-const tailwindcss = require("tailwindcss");
-const autoprefixer = require("autoprefixer");
+import tailwindcss from "tailwindcss";
+import autoprefixer from "autoprefixer";
 
-const config = {
+/** @type {import('postcss-load-config').Config} */
+export default {
   plugins: [
     //Some plugins, like tailwindcss/nesting, need to run before Tailwind,
-    tailwindcss(),
+    tailwindcss,
     //But others, like autoprefixer, need to run after,
     autoprefixer,
   ],
 };
-
-module.exports = config;
