@@ -1,5 +1,12 @@
 <script>
-  import "../app.pcss";
+  import "../app.css";
+  /**
+   * @typedef {Object} Props
+   * @property {import('svelte').Snippet} [children]
+   */
+
+  /** @type {Props} */
+  let { children } = $props();
 </script>
 
 <svelte:head>
@@ -10,13 +17,13 @@
   </style>
 </svelte:head>
 
-<slot />
+{@render children?.()}
 
 <footer
-  class="footer footer-center p-4 bg-base-300 text-base-content bottom-0 fixed"
+  class="footer footer-center fixed bottom-0 bg-base-300 p-4 text-base-content"
 >
   <aside>
-    <div class="flex flex-row max-w-lg w-full">
+    <div class="flex w-full max-w-lg flex-row">
       <a
         href="/donate"
         target="_blank"
