@@ -1,9 +1,9 @@
 <div class="flex">
   <div class="hero min-h-screen">
     <div class="hero-content text-center">
-      <div class="bg-base-200 p-16 max-w-lg rounded-md">
+      <div class="max-w-lg rounded-md bg-base-200 p-16">
         <h1 class="text-5xl font-bold">There was an error.</h1>
-        <div class="py-6 join">
+        <div class="join py-6">
           <a class="btn btn-primary btn-lg" href="/">Home</a>
         </div>
       </div>
@@ -11,7 +11,7 @@
   </div>
 
   <footer
-    class="footer footer-center p-4 bg-base-300 text-base-content bottom-0 fixed"
+    class="footer footer-center fixed bottom-0 bg-base-300 p-4 text-base-content"
   >
     <aside class="text-2xl">
       <p>
