@@ -2,7 +2,7 @@ import { vitePreprocess } from "@sveltejs/vite-plugin-svelte";
 import adapter from "@sveltejs/adapter-static";
 
 /** @type {import('@sveltejs/kit').Config} */
-const config = {
+export default {
   kit: {
     // adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
     // If your environment is not supported or you settled on a specific environment, switch out the adapter.
@@ -11,6 +11,7 @@ const config = {
   },
 
   preprocess: [vitePreprocess({})],
+  vitePlugin: {
+    inspector: true,
+  },
 };
-
-export default config;
